community/resume-builder/requirements.txt
python-docx>=0.8.11
reportlab>=3.6.12
